Fix an issue in envoy-fork
The wrong version of envoy-fork was being used. Specifically, the existing 4d15
commit had the 95363ffcc14059b279b6b618817aefd4bee1246b commit merged in, but
this is not currently used on envoy-gloo-ee. The new 2d00 commit performs the
tap filter cherry-pick again, but this time without that problematic 9536
commit. Hopefully this will get the chain of dependencies correct this time.