Update to v1.30 (#333)
* Update envoy dependency to v1.30.1

* Change repository back to upstream envoy

* Remove wavm

This was removed from upstream - see https://github.com/envoyproxy/envoy/pull/32872

* Add some compile flags to get CEL to compile

These flags are copied from upstream's .bazelrc. The most important one here is
the `-fsized-deallocation` which was required to get CEL to compile. The rest
may not be necessary, but since they changed in upstream from 1.29 to 1.30, I
think it might be a good idea to include them.

* Get matchers working

There were some details in regex that were changed upstream that we had to
incorporate

* Fix a few flags that were moved

* Get HTTP transformation to compile

A few upstream Envoy functions that were being called in this library needed to
be passed contexts, which required plumbing that object through the call
hierarchy. Along the way, it seemed to appear that there were a few type
signatures that were possibly not completely correct, and those have been
adjusted to get things compiling successfully.

* Fix some compile errors

It seems that passing string arguments constructed with the `+` operator is not
allowed in ENVOY_STREAM_LOG, and possibly other macros as well. This was my
first encounter with this error, but there might be other instances lurking
elsewhere in the repository.

* addWatch callback must now return absl::Status

* Fix type signature of createProtocolOptionsConfig

The signature of this abstract class method was changed upstream - this change
reflects the fix

* Add changelog

* Update extensions_build_config.bzl

* Move changelog

* Get inja_transformer_test.cc to compile

Interesting lesson here - when calling `fmt::format`, we must use a constexpr,
since the formatting function provides compile-time type checking. Neat!

* Fix a broken test

* Add router_check tool

Upstream ci calls this, so we put in a dummy target here

* Update bazel version

* do_ci.sh debug flags

* Update envoy-build-ubuntu image

* Remove debug lines from do_ci.sh

* Remove envoy.string_matcher.lua

Co-authored-by: Nathan Fudenberg <nathan.fudenberg@solo.io>

* Remove envoy.tracers.opentelemetry.samplers.dynatrace from bazel/extensions/extensions_build_config.bzl

Co-authored-by: Nathan Fudenberg <nathan.fudenberg@solo.io>

* Update changelog/v1.30.1-patch1/update-to-upstream-envoy-v1.30.yaml

Co-authored-by: Nathan Fudenberg <nathan.fudenberg@solo.io>

* Update source/extensions/filters/http/aws_lambda/config.cc

Co-authored-by: Nathan Fudenberg <nathan.fudenberg@solo.io>

* Revert "Update source/extensions/filters/http/aws_lambda/config.cc"

This reverts commit f91e83746aadc263ceb8136cd3f726590c538613.

* Fix one more review comment

* Revert `fmt::vformat` change in inja_transformer_test.cc

* Add comments explaining why `//test/tools` exist

* CHange auto variables to `std::string`

* Change `fmt::vformat` back to `fmt::format`

Previous iterations were still using the `vformat` call so they may not have
been using the `constexpr` versino of formatting correctly

---------

Co-authored-by: Nathan Fudenberg <nathan.fudenberg@solo.io>