support data source in transform (#397)
This PR introduces 2 new features for transformation.

First is the trim option which will trim ascii whitespace from both the start and end of the value provided.
This can be used like so: {{ trim(my_super_cool_value) }}

Second is the datasource option. This pulls from a mounted datasource to provide data or an empty string.
This can be used like so {{ data_source("my_cool_mounted_file") }}
One may often want to use the trim functionality in combination with this new feature.